move cmd/zqd to ppl/cmd/zqd
