+++ Makefile
@@ -89,10 +89,10 @@ zng-output-check: build $(SAMPLEDATA)
 # If the build recipe changes, please also change npm/build.
 build:
 	@mkdir -p dist
-	@go build -ldflags='$(LDFLAGS)' -o dist ./cmd/...
+	@go build -ldflags='$(LDFLAGS)' -o dist ./cmd/... ./ppl/cmd/...
 
 install:
-	@go install -ldflags='$(LDFLAGS)' ./cmd/...
+	@go install -ldflags='$(LDFLAGS)' ./cmd/... ./ppl/cmd/...
 
 docker:
 	DOCKER_BUILDKIT=1 docker build --pull --rm \
