+++ cmd/zapi/README.md
@@ -1,7 +1,7 @@
 # `zapi`
 
 `zapi` is command-line utility for interacting with the API provided by the
- [zqd](../zqd/README.md) daemon. It can be used to create data spaces,
+ [zqd](../../ppl/cmd/zqd/README.md) daemon. It can be used to create data spaces,
  import logs and pcaps, list info on spaces, and more.
 
 For all `zapi` options, use the help subcommand:
