+++ ppl/cmd/zqd/listen/command.go
@@ -17,12 +17,12 @@ import (
 	"strings"
 
 	"github.com/brimsec/zq/cli"
-	"github.com/brimsec/zq/cmd/zqd/logger"
-	"github.com/brimsec/zq/cmd/zqd/root"
 	"github.com/brimsec/zq/driver"
 	"github.com/brimsec/zq/pkg/fs"
 	"github.com/brimsec/zq/pkg/httpd"
 	"github.com/brimsec/zq/pkg/rlimit"
+	"github.com/brimsec/zq/ppl/cmd/zqd/logger"
+	"github.com/brimsec/zq/ppl/cmd/zqd/root"
 	"github.com/brimsec/zq/proc/sort"
 	"github.com/brimsec/zq/zqd"
 	"github.com/brimsec/zq/zqd/pcapanalyzer"
