+++ ppl/cmd/zqd/logger/waterfall.go
