+++ ppl/cmd/zqd/root/command.go
