+++ ppl/cmd/zqd/winexec/command.go
@@ -3,7 +3,7 @@ package winexec
 import (
 	"flag"
 
-	"github.com/brimsec/zq/cmd/zqd/root"
+	"github.com/brimsec/zq/ppl/cmd/zqd/root"
 	"github.com/mccanne/charm"
 )
