+++ zng/docs/spec.md
@@ -222,7 +222,7 @@ any ZNG streaming API.  In this way, senders and receivers of ZNG can embed
 protocol directives as ZNG control payloads rather than defining additional
 encapsulating protocols.
 
-> For example, the [zqd](../../cmd/zqd) server uses application-defined
+> For example, the [zqd](../../ppl/cmd/zqd) server uses application-defined
 > message `0xf7` to embed search and server stats in the return stream
 > of ZNG data, e.g., as a long-running search progresses on the server.
